moved a few controllers into feature folders, and split them up
