+++ src/Controller/Grocer/BuyController.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  */
 
 
-namespace App\Controller;
+namespace App\Controller\Grocer;
 
 use App\Entity\Inventory;
 use App\Entity\User;
@@ -35,33 +35,8 @@ use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/grocer')]
-class GrocerController extends AbstractController
+class BuyController extends AbstractController
 {
-    #[Route('', methods: ['GET'])]
-    #[IsGranted("IS_AUTHENTICATED_FULLY")]
-    public function getInventory(
-        GrocerService $grocerService, ResponseService $responseService, EntityManagerInterface $em
-    ): JsonResponse
-    {
-        /** @var User $user */
-        $user = $this->getUser();
-        $now = new \DateTimeImmutable();
-
-        $grocerItemsQuantity = UserQuestRepository::findOrCreate($em, $user, 'Grocer Items Purchased Quantity', 0);
-        $grocerItemsDay = UserQuestRepository::findOrCreate($em, $user, 'Grocer Items Purchased Date', $now->format('Y-m-d'));
-
-        if($now->format('Y-m-d') === $grocerItemsDay->getValue())
-            $maxCanPurchase = GrocerService::MAX_CAN_PURCHASE_PER_DAY - $grocerItemsQuantity->getValue();
-        else
-            $maxCanPurchase = GrocerService::MAX_CAN_PURCHASE_PER_DAY;
-
-        return $responseService->success([
-            'inventory' => $grocerService->getInventory(),
-            'maxPerDay' => GrocerService::MAX_CAN_PURCHASE_PER_DAY,
-            'maxRemainingToday' => $maxCanPurchase,
-        ]);
-    }
-
     #[Route('/buy', methods: ['POST'])]
     #[IsGranted("IS_AUTHENTICATED_FULLY")]
     public function buy(
