+++ src/Controller/Grocer/GetInventoryController.php
@@ -1,0 +1,54 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * This file is part of the Poppy Seed Pets API.
+ *
+ * The Poppy Seed Pets API is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ * The Poppy Seed Pets API is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with The Poppy Seed Pets API. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+
+namespace App\Controller\Grocer;
+
+use App\Entity\User;
+use App\Functions\UserQuestRepository;
+use App\Service\GrocerService;
+use App\Service\ResponseService;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+
+#[Route('/grocer')]
+class GetInventoryController extends AbstractController
+{
+    #[Route('', methods: ['GET'])]
+    #[IsGranted("IS_AUTHENTICATED_FULLY")]
+    public function getInventory(
+        GrocerService $grocerService, ResponseService $responseService, EntityManagerInterface $em
+    ): JsonResponse
+    {
+        /** @var User $user */
+        $user = $this->getUser();
+        $now = new \DateTimeImmutable();
+
+        $grocerItemsQuantity = UserQuestRepository::findOrCreate($em, $user, 'Grocer Items Purchased Quantity', 0);
+        $grocerItemsDay = UserQuestRepository::findOrCreate($em, $user, 'Grocer Items Purchased Date', $now->format('Y-m-d'));
+
+        if($now->format('Y-m-d') === $grocerItemsDay->getValue())
+            $maxCanPurchase = GrocerService::MAX_CAN_PURCHASE_PER_DAY - $grocerItemsQuantity->getValue();
+        else
+            $maxCanPurchase = GrocerService::MAX_CAN_PURCHASE_PER_DAY;
+
+        return $responseService->success([
+            'inventory' => $grocerService->getInventory(),
+            'maxPerDay' => GrocerService::MAX_CAN_PURCHASE_PER_DAY,
+            'maxRemainingToday' => $maxCanPurchase,
+        ]);
+    }
+}
