+++ src/Controller/Guilds/GetAllController.php
@@ -12,46 +12,22 @@ declare(strict_types=1);
  */
 
 
-namespace App\Controller;
+namespace App\Controller\Guilds;
 
 use App\Entity\Guild;
 use App\Enum\SerializationGroupEnum;
-use App\Service\Filter\GuildMemberFilterService;
 use App\Service\ResponseService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
 
 #[Route("/guild")]
-class GuildController extends AbstractController
+class GetAllController extends AbstractController
 {
     #[Route("", methods: ["GET"])]
     public function getAll(ResponseService $responseService, EntityManagerInterface $em): JsonResponse
     {
         return $responseService->success($em->getRepository(Guild::class)->findAll(), [ SerializationGroupEnum::GUILD_ENCYCLOPEDIA ]);
-    }
-
-    #[Route("/{guild}", methods: ["GET"])]
-    public function getGuild(
-        Guild $guild, ResponseService $responseService, GuildMemberFilterService $guildMemberFilterService, Request $request
-    ): JsonResponse
-    {
-        $guildMemberFilterService->addRequiredFilter('guild', $guild->getId());
-
-        $members = $guildMemberFilterService->getResults($request->query);
-
-        return $responseService->success(
-            [
-                'guild' => $guild,
-                'members' => $members
-            ],
-            [
-                SerializationGroupEnum::FILTER_RESULTS,
-                SerializationGroupEnum::GUILD_MEMBER,
-                SerializationGroupEnum::GUILD_ENCYCLOPEDIA,
-            ]
-        );
     }
 }
