+++ src/Controller/Letters/ReadController.php
@@ -12,45 +12,22 @@ declare(strict_types=1);
  */
 
 
-namespace App\Controller;
+namespace App\Controller\Letters;
 
 use App\Entity\User;
 use App\Entity\UserLetter;
-use App\Enum\SerializationGroupEnum;
 use App\Exceptions\PSPNotFoundException;
 use App\Service\FieldGuideService;
-use App\Service\Filter\UserLetterFilterService;
 use App\Service\ResponseService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route("/letter")]
-class LetterController extends AbstractController
+class ReadController extends AbstractController
 {
-    #[Route("", methods: ["GET"])]
-    #[IsGranted("IS_AUTHENTICATED_FULLY")]
-    public function getLetters(
-        Request $request, ResponseService $responseService,
-        UserLetterFilterService $userLetterFilterService
-    ): JsonResponse
-    {
-        /** @var User $user */
-        $user = $this->getUser();
-
-        $userLetterFilterService->addRequiredFilter('user', $user->getId());
-
-        $results = $userLetterFilterService->getResults($request->request);
-
-        return $responseService->success($results, [
-            SerializationGroupEnum::FILTER_RESULTS,
-            SerializationGroupEnum::MY_LETTERS
-        ]);
-    }
-
     #[Route("/{letter}/read", methods: ["PATCH"])]
     #[IsGranted("IS_AUTHENTICATED_FULLY")]
     public function markRead(
